feat: Add functionality to create a "signtool.exe" (#7)
This PR adds functionality to have `@electron/windows-sign` create a
`signtool.exe` that signs any given file with the same options passed to
`@electron/windows-sign` in the first place. That's helpful for tools
like Squirrel - which will _only_ sign with `signtool.exe`. It's a bit
of a clever hack to make `@electron/windows-sign` compatible with any
tool that'll let users call `signtool.exe`.

Here's how it works:
- On the user's machine, we use Node's "single executable binary"
feature to create a custom version of `node.exe` that contains signing
options. We'll call that newly minted binary `signtool.exe`.
- When our `signtool.exe` is executed, it spawns user's `node.exe`,
requires `@electron/windows-sign`, and signs the file originally passed
to it with the options already embedded in the binary.
- Since we _create_ the `signtool.exe` on the local machine, users can
use this tool with confidence - there is no mysterious "signtool.exe"
that'd be able to exfiltrate secrets.

I don't expect users to actually use this feature directly. It'll be
consumed by `@electron/windows-installer`, for which I'll open another
PR.