fix: SEA should not throw unless script errors on exit (#57)
Depending on the shell used (like PowerShell), a lot of output will be
auto-piped to `stderr` - even if it's not an error at all. That has
entire build pipelines failing even though there is no error at all.

To fix this, we'll look at the exit code of the invoked script. If the
process exited with anything other than `0`, we'll error and bubble up
`stderr`.

We didn't notice that earlier because I am an insane person using git
bash every now and then