chore: clean up formatting setup (#60)
Standardizes this repo's formatting setup around Prettier instead of
Standard + ESLint.

There was already a `.prettierrc` file included in the repo, which was
causing Format on Save to throw lint errors in development.

This PR fixes that, simplifies the ESLint config, updates
`@typescript-eslint` deps to match the TS version we're using, and adds
pre-commit hooks with lint-staged + husky.